TASK: Apply new core migrations
Expand Neos.Neos-FusionObjects without namespace to fully qualified
names

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration Neos.Fusion-20180211184832

Remove namespace aliases and replace them with the package keys

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

TASK: Apply migration Neos.Flow-20180415105700

Add scalar type hint to CacheAwareInterface implementations.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.